requirements.txt
anthropic==0.45.0
openai==0.28.1
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers
torch
transformers
